+++ src/mkvtoolnix-gui/util/settings.cpp
@@ -601,7 +601,6 @@ Settings::load() {
   m_uiLocale                                  = reg.value(s_valUiLocale).toString();
 #endif
   m_uiDisableHighDPIScaling                   = reg.value(s_valUiDisableHighDPIScaling).toBool();
-  m_uiDisableDarkStyleSheet                   = reg.value(s_valUiDisableDarkStyleSheet).toBool();
   m_uiDisableToolTips                         = reg.value(s_valUiDisableToolTips).toBool();
   m_uiFontFamily                              = reg.value(s_valUiFontFamily,                                                                   defaultFont.family()).toString();
   m_uiFontPointSize                           = reg.value(s_valUiFontPointSize,                                                                defaultFont.pointSize()).toInt();
@@ -1036,7 +1035,6 @@ Settings::save()
 
   reg.setValue(s_valUiLocale,                                  m_uiLocale);
   reg.setValue(s_valUiDisableHighDPIScaling,                   m_uiDisableHighDPIScaling);
-  reg.setValue(s_valUiDisableDarkStyleSheet,                   m_uiDisableDarkStyleSheet);
   reg.setValue(s_valUiDisableToolTips,                         m_uiDisableToolTips);
   reg.setValue(s_valUiFontFamily,                              m_uiFontFamily);
   reg.setValue(s_valUiFontPointSize,                           m_uiFontPointSize);